change name for static data file, extract extra work into function.
