+++ src/App.jsx
@@ -1,7 +1,9 @@
 import { useState, useEffect } from "react";
 import RowContainer from "./components/RowContainer";
 import Stats from "./components/Stats.jsx";
-import defaultRows from "./data/row.jsx";
+import defaultRows from "./data/rowData.jsx";
+import sortRows from "./utils/sortRows.jsx";
+import checkIsSameOrder from "./utils/checkIsSameOrder.jsx";
 
 function App() {
   const [rows, setRows] = useState(defaultRows);
@@ -9,13 +11,9 @@ function App() {
 
   useEffect(() => {
     // Sort rows whenever any counterValue changes
-    const sortedRows = [...rows].sort(
-      (a, b) => a.counterValue - b.counterValue
-    );
+    const sortedRows = sortRows(rows);
 
-    const isSameOrder = rows.every(
-      (item, index) => item.id === sortedRows[index]?.id
-    );
+    const isSameOrder = checkIsSameOrder(rows, sortedRows);
     // if the order is same then we don't change the rows
     //this will also prevent it from going to infinite re-rendering
     if (!isSameOrder) {
