+++ src/utils/checkIsSameOrder.jsx
@@ -1,0 +1,3 @@
+export default function checkIsSameOrder(origRows, sortedRows) {
+  return origRows.every((item, index) => item.id === sortedRows[index]?.id);
+}
