+++ src/utils/sortRows.jsx
@@ -1,0 +1,3 @@
+export default function sortRows(rows) {
+  return [...rows].sort((a, b) => a.counterValue - b.counterValue);
+}
